clearway-for-gmail/pyproject.toml
[project]
name = "clearway-for-gmail"
version = "0.1.0"
description = "Clearway bridge for Gmail using Google API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.115.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.1",
]
